user authorisation check added to delete endpoint

+++ part4/bloglist/controllers/blogs.js
@@ -51,13 +51,28 @@ blogRouter.post('/', async (request, response, next) => {
 })
 
 blogRouter.delete('/:id', async (request, response, next) => {
+
   try {
+    const token = request.token
+    if (!token) {
+      response.status(400).json({ error: 'Token missing' })
+    }
+
+    const decodedToken = jwt.decode(token)
+    const userID = decodedToken.id
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() !== userID) {
+      return response.status(400).json({ error: 'Invalid user' })
+    }
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
+
   } catch (error) {
-    next(error)
-  }
+    next(error) }
 })
+
 
 blogRouter.put('/:id', async (request, response, next) => {
   const likes = request.body.likes
